refactor(server): move agent edit runtimes into package
