--- packages/doc-server/src/agent-edit-ops-shared.ts
@@ -1,6 +1,0 @@
-export { applyAgentEditOperations } from '../../../server/agent-edit-ops.js';
-
-export type {
-  AgentEditApplyResult,
-  AgentEditOperation,
-} from '../../../server/agent-edit-ops.js';
