--- packages/doc-server/src/agent-edit-v2-shared.ts
@@ -1,3 +1,0 @@
-export { applyAgentEditV2 } from '../../../server/agent-edit-v2.js';
-
-export type { AgentEditV2Result } from '../../../server/agent-edit-v2.js';
