+++ packages/doc-server/src/canonical-document-shared.ts
@@ -1,4 +1,5 @@
 export {
+  deriveProjectionFromCanonicalDoc,
   cloneFromCanonical,
   executeCanonicalRewrite,
   mutateCanonicalDocument,
