+++ src/tests/server-package-internals.test.ts
@@ -9,9 +9,9 @@ async function run(): Promise<void> {
   const shimCanonical = await import('../../server/canonical-document.ts');
   const packageEngine = await import('../../packages/doc-server/src/document-engine-shared.ts');
   const shimEngine = await import('../../server/document-engine.ts');
-  const packageAgentEditOps = await import('../../packages/doc-server/src/agent-edit-ops-shared.ts');
+  const packageAgentEditOps = await import('../../packages/doc-server/src/agent-edit-ops.ts');
   const shimAgentEditOps = await import('../../server/agent-edit-ops.ts');
-  const packageAgentEditV2 = await import('../../packages/doc-server/src/agent-edit-v2-shared.ts');
+  const packageAgentEditV2 = await import('../../packages/doc-server/src/agent-edit-v2.ts');
   const shimAgentEditV2 = await import('../../server/agent-edit-v2.ts');
   const packageHeadless = await import('../../packages/doc-server/src/milkdown-headless.ts');
   const shimHeadless = await import('../../server/milkdown-headless.ts');
@@ -49,11 +49,11 @@ async function run(): Promise<void> {
   }
 
   if (packageAgentEditOps.applyAgentEditOperations !== shimAgentEditOps.applyAgentEditOperations) {
-    throw new Error('Expected agent edit ops facade to preserve applyAgentEditOperations');
+    throw new Error('Expected agent edit ops shim to preserve applyAgentEditOperations');
   }
 
   if (packageAgentEditV2.applyAgentEditV2 !== shimAgentEditV2.applyAgentEditV2) {
-    throw new Error('Expected agent edit v2 facade to preserve applyAgentEditV2');
+    throw new Error('Expected agent edit v2 shim to preserve applyAgentEditV2');
   }
 
   if (packageHeadless.getHeadlessMilkdownParser !== shimHeadless.getHeadlessMilkdownParser) {
